% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmPR.R
\name{glmPR.default}
\alias{glmPR.default}
\title{L1-penalized Poisson Regression default function}
\usage{
\method{glmPR}{default}(X, y, lambda = 1, threads = 4, ...)
}
\arguments{
\item{X}{the input matrix}

\item{y}{the response vector}

\item{lambda}{a constant scalar parameter to control the influence of L1-Norm}

\item{threads}{the parallelize node number}

\item{...}{other parameter}
}
\value{
coefficients vector
}
\description{
L1-penalized Poisson Regression default function
}
\examples{
x <- matrix(rnorm(100), ncol = 4)
y <- rpois(25, 3)
glmPR(x, y, 1.0, 4)

}
