% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmPR.R
\name{glmPR}
\alias{glmPR}
\title{L1-penalized Poisson Regression
You can see glmPR.default and glmPR.formula for more details}
\usage{
glmPR(X, ...)
}
\arguments{
\item{X}{the input matrix}

\item{...}{other parameter}
}
\value{
coefficients vector
}
\description{
L1-penalized Poisson Regression
You can see glmPR.default and glmPR.formula for more details
}
