% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmPR.R
\name{predict.glmPR}
\alias{predict.glmPR}
\title{The predict funtion with new data}
\usage{
\method{predict}{glmPR}(object, newx = NULL, ...)
}
\arguments{
\item{object}{the learned glmPR object}

\item{newx}{the new itest data}

\item{...}{other parameter}
}
\value{
the predicted vector
}
\description{
The predict funtion with new data
}
\examples{
x <- matrix(rnorm(100), ncol = 4)
y <- rpois(25, 3)
fit <- glmPR(x, y)
predict(fit, x[1:3,])

}
